fix: use wait-groups for channel reading
Because you cannot guarantee what order go-routines are executed in,
sometimes the order the repos were parsed in didn't line up with the
order the channel was read from. When this happens, you get confusing
results in debug mode as the repo often doesn't match the message that
is given back.