feat: add templating to arch and os config fields
Allow the "arch" and "os" fields in the config file to be templated.
This is a bit of a hack considering that we have to override the Data
Map with the default GOOS and GOARCH in order to get it to work. But it
allows users to override specific fields instead of having to specify
the entire ExternalURL if their project of choice does something odd
with the Arch or OS fields.

Examples I've used it for:

```
  # A k9s-esque application for viewing Cloud resources from AWS or GCP
  - repo: one2nc/cloudlens
    arch: '{{ if eq .os "darwin" }}all{{ else }}{{ .arch }}{{ end }}'
  # A slightly more robust and easy to use tool than sed for replacing text in files
  - repo: your-tools/ruplacer
    os: '{{ if eq .os "darwin" }}apple-darwin{{ else if eq .os "linux" }}unknown-linux{{ else }}pc-windows{{ end }}'
    arch: '{{ if eq .os "darwin"}}aarch64{{ else }}{{ .arch }}{{ end }}'
```

Signed-off-by: Aaron U'Ren <aauren@gmail.com>